+++ lib/main.dart
@@ -11,38 +11,14 @@ import 'package:fb/providers/transaction.dart';
 import 'package:fb/quick_transaction.dart';
 import 'package:flutter/material.dart';
 import 'package:home_widget/home_widget.dart';
-import 'package:money2/money2.dart';
+import 'package:http/io_client.dart';
 import 'package:provider/provider.dart';
 import 'package:realm/realm.dart' as realm;
 
 import 'app.dart';
 
-// todo do refactoring to extract widgets and functions as much as possible (globally)
-
-// todo Obfuscate Dart code as part of CI/CD and https://codewithandrea.com/articles/flutter-api-keys-dart-define-env-files/
-
-// todo performance issues (https://docs.flutter.dev/perf/best-practices, devtools in profile mode, find other ways to optimize)
-
-// todo https://pub.dev/packages/dynamic_color
-
 Future<void> main() async {
-  // todo download and add font with license https://github.com/material-foundation/flutter-packages/blob/main/packages/google_fonts/README.md#bundling-fonts-when-releasing
-  /*LicenseRegistry.addLicense(() async* {
-    final license = await rootBundle.loadString('google_fonts/OFL.txt');
-    yield LicenseEntryWithLineBreaks(['google_fonts'], license);
-  });*/
-
   WidgetsFlutterBinding.ensureInitialized();
-  Currencies().registerList(<Currency>[
-    Currency.create('UAH', 2,
-        symbol: '₴',
-        name: 'Ukrainian Hryvnia',
-        country: 'Ukraine',
-        pattern: 'S0.00',
-        unit: 'hryvnia'),
-    Currency.create('AED', 2,
-        symbol: 'د.إ', name: 'The United Arab Emirates dirham', country: 'The United Arab Emirate', unit: 'dirham'),
-  ]);
 
   HomeWidget.initiallyLaunchedFromHomeWidget().then((value) {
     if (value != null) {
@@ -54,7 +30,7 @@ Future<void> main() async {
 }
 
 Future<void> _runApp(Widget app) async {
-  final realmApp = realm.App(realm.AppConfiguration(GlobalConfig().realmAppID, httpClient: HttpClient()));
+  final realmApp = realm.App(realm.AppConfiguration(GlobalConfig().realmAppID, httpClient: IOClient(HttpClient())));
   app = const App();
 
   StateProvider stateProvider = StateProvider(realmApp);