+++ lib/providers/category.dart
@@ -1,8 +1,8 @@
 import 'package:fb/db/category.dart';
 import 'package:fb/db/repository.dart';
 import 'package:fb/models/category.dart';
+import 'package:fb/utils/currency.dart';
 import 'package:flutter/material.dart';
-import 'package:money2/money2.dart';
 
 class CategoryProvider extends ChangeNotifier {
   List<Category> _categories = [];